- Ongoing IdGenerator refactoring

+++ ums-api/src/main/java/org/ums/domain/model/immutable/SeatPlan.java
@@ -11,7 +11,7 @@ import java.io.Serializable;
  * Created by My Pc on 5/8/2016.
  */
 public interface SeatPlan extends Serializable, LastModifier, EditType<MutableSeatPlan>,
-    Identifier<Integer> {
+    Identifier<Long> {
   ClassRoom getClassRoom();
 
   Student getStudent();
